Add special case to `class` attribute for `Svg.Attributes.class`
Currently, `Html.Attributes.class` is special cased so that it can be
provided
multiple times, and the entries will be appended rather than overwriting
one
another.

It is rather surprising to the user that this behaves differently when
dealing
with `Svg`.

This change is slightly more complicated due to svg requiring the
`class`
attribute rather than a property.