Minimise white flash between deleting and replacing nodes in DOM
This improves the visual smoothness of hand-over between static
rendered HTML and dynamic HTML that is produced by Elm.

Currently, if we have a large amount of statically generated HTML
which is replaced by a large amount of HTML rendered by Elm, on
running `App.Main.embed` there will be noticeable period between
the old static content getting removed and the new content being
added. Depending on the styling this can result in the appearance
of a "white flash".

This commit does the absolute minimum to bring the removal and addition
closer together so that the white flash is minimised. There are further
optimisations that could be done. For instance, we could in the
case of a single child use `node.replaceChild` instead of
`node.removeChild` followed by `node.appendChild`.