Fix which `key` is used in swapping nodes
Consider this:
x is `A B C D`
y is `A C B D`

B is a new match and the xNode and yNextNode
C is an old match and the yNode and xNextNode

yNode == xNode

We first diff x's B and y's B `_VirtualDom_diffHelp(xNode, yNextNode, localPatches, index);`
Then we insert y's C (yNode)
Then we remove the x's C (xNextNode)

When we do the insert/remove of C we should be using yKey (or xNextKey) because that is what C's key is.

With all that being said,

This is only a semantic bug. The code works fine because since xKey is used in both places it becomes a Move like it would otherwise.