Use Object.create(null) instead of {} to compute keyed diffs
`{}` creates an empty object but which still contains methods and
properties like `toString`. Checking `changes[key]` when the key is
for instance `toString` creates false positives in that check.

In practice, the algorithms in `_VirtualDom_insertNode` and `_VirtualDom_removeNode`
end up with the correct behavior because they recurse with a modified
key, but do so with an unnecessary iteration.

`Object.create(null)` create an empty object without any methods or
properties, leading to no false positives.